+++ APILambda/package.json
@@ -23,7 +23,7 @@
     "linkify-it": "^5.0.0",
     "TiFBackendUtils": "file:../TiFBackendUtils",
     "uuid": "^11.0.2",
-    "TiFShared": "github:tifapp/TiFShared#e0445304b365d4c5a38c156fc7b4f1f7136273ee",
+    "TiFShared": "github:tifapp/TiFShared#ae5cd19143c31e380edaca2b04fc64f1dc6dd7b8",
     "zod": "^3.22.4"
   },
   "devDependencies": {
@@ -41,7 +41,6 @@
     "@types/node-fetch": "^2.6.4",
     "@types/supertest": "^2.0.12",
     "concurrently": "^8.0.1",
-    "dayjs": "^1.11.10",
     "expect-type": "^0.18.0",
     "jest": "^29.7.0",
     "jest-watch-suspend": "^1.1.2",
