+++ APILambda/test/location.ts
@@ -1,3 +1,4 @@
+import { faker } from "@faker-js/faker"
 import { conn } from "TiFBackendUtils"
 import { LocationCoordinate2D } from "TiFShared/domain-models/LocationCoordinate2D"
 import { addLocationToDB } from "../../GeocodingLambda/utils"
@@ -15,3 +16,15 @@ export const addMockLocationToDB = (coords: LocationCoordinate2D) =>
     },
     "Sample/Timezone"
   )
+
+export const geocodeMock = async (coords: LocationCoordinate2D) => ({
+  ...coords,
+  name: faker.company.name(),
+  country: faker.address.country(),
+  postalCode: faker.address.zipCode(),
+  street: faker.address.street(),
+  streetNumber: faker.address.buildingNumber(),
+  region: faker.address.state(),
+  isoCountryCode: faker.address.countryCode(),
+  city: faker.address.city()
+})
